genai/Scripts/requirements.txt
boto3>=1.28.59
python-dotenv
spacy
googlesearch-python
wikipedia
duckduckgo-search
wikipedia-api
streamlit
streamlit_chat
langchain
sentence_transformers
tiktoken
openai
python-dotenv
huggingface_hub
pinecone-client
unstructured 
boto3
botocore
PyMuPDF
python-docx
llmsherpa
trp
amazon-textract-prettyprinter
amazon-textract-caller
textract-trp
opencv-python-headless
amazon-textract-textractor[pdf]
amazon-textract-textractor
IPython
googlesearch-python
tabula-py
streamlit
streamlit_chat
langchain
sentence_transformers
tiktoken
openai
python-dotenv
huggingface_hub
pinecone-client
unstructured 
boto3
botocore
PyMuPDF
python-docx
llmsherpa
faiss-cpu 
transformers
Pillow
pypdf